+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased] - ~
 ### Fixed
  - Mark debug output as debug
+ - Tooltip scanning on re-logging
 
 ## [0.2.1] - 2021-03-24
 ### Fixed